Adding Bridge module to handle admin shapes; adding deployment plan
