--- Constants/FeatureNames.cs
@@ -1,8 +1,0 @@
-﻿namespace Lombiq.Hosting.MediaTheme.Constants;
-
-public static class FeatureNames
-{
-    public const string Module = "Lombiq.Hosting.MediaTheme";
-
-    public const string MediaTheme = Module;
-}
