--- Constants/Paths.cs
@@ -1,9 +1,0 @@
-﻿namespace Lombiq.Hosting.MediaTheme.Constants;
-
-public static class Paths
-{
-    public const string MediaThemeRootPath = "_MediaTheme";
-    public const string MediaThemeTemplatesFolder = "Views";
-    public const string MediaThemeAssetsFolder = "Assets";
-    public const string MediaThemeAssetsPath = MediaThemeRootPath + "/" + MediaThemeAssetsFolder;
-}
