--- Constants/Routes.cs
@@ -1,6 +1,0 @@
-﻿namespace Lombiq.Hosting.MediaTheme.Constants;
-
-public static class Routes
-{
-    public const string MediaThemeAssets = "/mediatheme";
-}
