+++ Lombiq.Hosting.MediaTheme.Bridge/Constants/FeatureNames.cs
@@ -1,0 +1,9 @@
+﻿namespace Lombiq.Hosting.MediaTheme.Bridge.Constants;
+
+public static class FeatureNames
+{
+    private const string Base = "Lombiq.Hosting.MediaTheme";
+
+    public const string MediaTheme = Base;
+    public const string MediaThemeBridge = MediaTheme + ".Bridge";
+}
