+++ Lombiq.Hosting.MediaTheme.Bridge/Constants/Paths.cs
@@ -1,0 +1,9 @@
+﻿namespace Lombiq.Hosting.MediaTheme.Bridge.Constants;
+
+public static class Paths
+{
+    public const string MediaThemeRootFolder = "_MediaTheme";
+    public const string MediaThemeTemplatesFolder = "Views";
+    public const string MediaThemeAssetsFolder = "wwwroot";
+    public const string MediaThemeAssetsWebPath = MediaThemeRootFolder + "/" + MediaThemeAssetsFolder;
+}
