+++ Lombiq.Hosting.MediaTheme.Bridge/Constants/RecipeStepIds.cs
@@ -1,0 +1,6 @@
+﻿namespace Lombiq.Hosting.MediaTheme.Bridge.Constants;
+
+public static class RecipeStepIds
+{
+    public const string MediaTheme = "mediatheme";
+}
