+++ Lombiq.Hosting.MediaTheme.Bridge/Constants/Routes.cs
@@ -1,0 +1,6 @@
+﻿namespace Lombiq.Hosting.MediaTheme.Bridge.Constants;
+
+public static class Routes
+{
+    public const string MediaThemeAssets = "/mediatheme";
+}
