+++ Lombiq.Hosting.MediaTheme.Bridge/Models/MediaTemplate.cs
@@ -1,0 +1,10 @@
+﻿using Microsoft.AspNetCore.Html;
+
+namespace Lombiq.Hosting.MediaTheme.Bridge.Models;
+
+public class MediaTemplate
+{
+    public string TemplatePath { get; set; }
+    public string ShapeType { get; set; }
+    public string Content { get; set; }
+}
