+++ Lombiq.Hosting.MediaTheme.Bridge/Models/MediaThemeStateDocument.cs
@@ -1,0 +1,8 @@
+﻿using OrchardCore.Data.Documents;
+
+namespace Lombiq.Hosting.MediaTheme.Bridge.Models;
+
+public class MediaThemeStateDocument : Document
+{
+    public string BaseThemeId { get; set; }
+}
