+++ Lombiq.Hosting.MediaTheme.Bridge/Services/IMediaThemeStateStore.cs
@@ -1,0 +1,25 @@
+﻿using Lombiq.Hosting.MediaTheme.Bridge.Models;
+using System.Threading.Tasks;
+
+namespace Lombiq.Hosting.MediaTheme.Bridge.Services;
+
+/// <summary>
+/// Manages the stored media theme state (e.g., base theme).
+/// </summary>
+public interface IMediaThemeStateStore
+{
+    /// <summary>
+    /// Loads the media theme store for editing.
+    /// </summary>
+    Task<MediaThemeStateDocument> LoadMediaThemeStateAsync();
+
+    /// <summary>
+    /// Fetches the media theme store from cache. Not suitable for updating.
+    /// </summary>
+    Task<MediaThemeStateDocument> GetMediaThemeStateAsync();
+
+    /// <summary>
+    /// Saves the updated state document.
+    /// </summary>
+    Task SaveMediaThemeStateAsync(MediaThemeStateDocument state);
+}
