+++ Lombiq.Hosting.MediaTheme.Bridge/ViewModels/MediaThemeDeploymentStepViewModel.cs
@@ -1,0 +1,8 @@
+using System.Collections.Generic;
+
+namespace Lombiq.Hosting.MediaTheme.Bridge.ViewModels;
+
+public class MediaThemeDeploymentStepViewModel
+{
+    public bool ClearMediaThemeFolder { get; set; }
+}
