+++ Lombiq.Hosting.MediaTheme.Bridge/ViewModels/MediaThemeSettingsViewModel.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.AspNetCore.Mvc.ModelBinding;
+using OrchardCore.DisplayManagement.Extensions;
+using System.Collections.Generic;
+
+namespace Lombiq.Hosting.MediaTheme.Bridge.ViewModels;
+
+public class MediaThemeSettingsViewModel
+{
+    [BindNever]
+    public IEnumerable<(string Id, string Name)> AvailableBaseThemes { get; set; }
+
+    public string BaseThemeId { get; set; }
+}
