+++ Lombiq.Hosting.MediaTheme.Bridge/Views/Items/MediaThemeDeploymentStep.Fields.Thumbnail.cshtml
@@ -1,0 +1,4 @@
+@model dynamic
+
+<h4 class="card-title">@T["Media Theme"]</h4>
+<p>@T["Exports media theme settings."]</p>
