+++ Lombiq.Hosting.MediaTheme.Bridge/Views/_ViewImports.cshtml
@@ -1,0 +1,11 @@
+@inherits OrchardCore.DisplayManagement.Razor.RazorPage<TModel>
+
+@using OrchardCore.ContentManagement.Display.ViewModels
+@using OrchardCore.DisplayManagement.Views
+@using Lombiq.Hosting.MediaTheme.Bridge.Models
+@using Lombiq.Hosting.MediaTheme.Bridge.ViewModels
+
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
+@addTagHelper *, OrchardCore.DisplayManagement
+@addTagHelper *, OrchardCore.ResourceManagement
+@addTagHelper *, OrchardCore.Contents
