--- Manifest.cs
@@ -1,11 +1,0 @@
-using OrchardCore.DisplayManagement.Manifest;
-
-[assembly: Theme(
-    Name = "Lombiq Hosting - Media Theme",
-    Author = "Lombiq Technologies",
-    Website = "https://github.com/Lombiq/Hosting-Tenants",
-    Version = "1.0.0",
-    Description = "Ability to host a theme from the Media Library.",
-    Category = "Hosting",
-    Dependencies = new[] { "OrchardCore.Media", "OrchardCore.Users" }
-)]
