--- Models/MediaTemplate.cs
@@ -1,10 +1,0 @@
-﻿using Microsoft.AspNetCore.Html;
-
-namespace Lombiq.Hosting.MediaTheme.Models;
-
-public class MediaTemplate
-{
-    public string TemplatePath { get; set; }
-    public string ShapeType { get; set; }
-    public string Content { get; set; }
-}
