--- Models/MediaThemeStateDocument.cs
@@ -1,8 +1,0 @@
-﻿using OrchardCore.Data.Documents;
-
-namespace Lombiq.Hosting.MediaTheme.Models;
-
-public class MediaThemeStateDocument : Document
-{
-    public string BaseThemeId { get; set; }
-}
