--- Views/_ViewImports.cshtml
@@ -1,11 +1,0 @@
-@inherits OrchardCore.DisplayManagement.Razor.RazorPage<TModel>
-
-@using OrchardCore.ContentManagement.Display.ViewModels
-@using OrchardCore.DisplayManagement.Views
-@using Lombiq.Hosting.MediaTheme.Models
-@using Lombiq.Hosting.MediaTheme.ViewModels
-
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-@addTagHelper *, OrchardCore.DisplayManagement
-@addTagHelper *, OrchardCore.ResourceManagement
-@addTagHelper *, OrchardCore.Contents
